Squashed commit:api structure changes
commit 0e9d90fe8d067f7669210611a4e6a9b463ea0d6e
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Mon May 10 20:03:14 2021 +0530

    add filter for employee

commit 3078c027845a05ab37d5e9a925edc3f77975f81d
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Mon May 10 15:53:00 2021 +0530

    fix document list query

commit 8c9441e6b86b71bea69b200f1df50b837dddb350
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Mon May 10 11:07:08 2021 +0530

    fix task list api query for client panel

commit 5a5b9a88fb340f3c9e903d430b0d2863ad46e2bb
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Mon May 10 10:58:11 2021 +0530

    fix task list api

commit 6fced7f69c5dc253fc6ed0ea349a6e9ebc2b177d
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Fri May 7 10:18:14 2021 +0530

    debtor api changes

commit b98ebfb399af850216c74e8b8a2a4843b191435b
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Thu May 6 15:13:18 2021 +0530

    fix debtor details api

commit 6638ca2de097860cfde161776c36767f80857c34
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Thu May 6 15:00:57 2021 +0530

    fix country name field for debtor

commit ed5759e3f0068600dcb9ba35633a73c377a15f71
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Thu May 6 12:14:00 2021 +0530

    add street type for debtor

commit f75f032b8bb622b7ad9736733fdb2a2213c8edb6
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Thu May 6 11:34:54 2021 +0530

    debtor entity list + doc type api changes

commit b5fcf5078efda24a22b6f2fff96427d217acb321
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Tue May 4 17:38:45 2021 +0530

    stakeholder response strusture changes

commit 7453cae62e8ce9f78adbab5d6fc7112777dca2c7
Author: Chandani Janani <chandani.janani@kevit.io>
Date:   Tue May 4 16:44:03 2021 +0530

    add ABN search apis for debtor

+++ routes/document.risk.route.js
@@ -487,7 +487,7 @@ router.post('/upload', upload.single('document'), async function (req, res) {
       uploadByType: 'user',
     });
     const entityName = await getEntityName({
-      entityId: req.body.entityRefId,
+      entityId: req.body.entityId,
       entityType: req.body.documentFor.toLowerCase(),
     });
     await addAuditLog({
@@ -588,7 +588,9 @@ router.delete('/:documentId', async function (req, res) {
     const document = await Document.findOne({
       _id: req.params.documentId,
     }).lean();
-    await deleteFile({ filePath: document.keyPath });
+    if (document.keyPath) {
+      await deleteFile({ filePath: document.keyPath });
+    }
     await Document.updateOne(
       { _id: req.params.documentId },
       { isDeleted: true },
